git: don't set authentication to false if no authentication was requsted: IDEA-221395
Since af4757b2 (IDEA-217990, IDEA-218129) all Git commands are being run under an authentication manager, which means that all of them listen to Git output and detect the "authentication failure" situation if certain text patterns are found in the output. Including, for instance, commit messages which occur in a git log output.

If such a pattern is detected, the myHttpAuthFailed flag is set, which may lead to unexpected behavior. For example, `git reset --hard` could be repeated 3 times in a row (it is not happening right now, but only because the command is called using another call chain which doesn't pass through the retrying logic). Also, all such "authentication failed" messages are written the log, which is confusing.

GitOrigin-RevId: a8b2cbdf91a6c08cd84fc9c7be76c7ca55a499f9