build scripts: ignore dependencies on wadl-core and jshell-frontend libraries while building Maven artifacts
This is needed to publish intellij.java.impl and other modules which depend on it as Maven artifacts (IDEA-220226). It's unlikely that code which depend on these libraries will be used when running tests so skipping these dependencies shouldn't cause real problems since for now Maven artifacts are used for compilation and running tests only.

The change can be reverted when IDEA-220227 and IDEA-222381 are fixed.

GitOrigin-RevId: 222ab1cb4876ac0b5787adf5e35a5fdc2f8bdeac