Make PyAncestorTypeProvider last in the list of type providers
It's a temporary workaround for the fact that for some reason explicitly stating
that NumpyDocStringTypeProvider should go *before* PyDocStringTypeProvider moves
the latter after the former, not vice-versa, thus making PyDocStringTypeProvider
be considered after all the other type providers, some of which intended to work
as a fallback source of the type information, including PyAncestorTypeProvider.

It makes it impossible to override ancestor's return type in a docstring,
say, in python-skeletons, as it was done for os.wrap_close.close method.

GitOrigin-RevId: 8ad4a0a908efc8ac3e6393a111fc76ac2d15a04d