until we don't have performance problems with preloading of services activity, simply await after component creation and do no complicate without a reason
if preloading of services will be not awaited, then service must subscribe to events using declarative way (because otherwise such events may be skipped), but declarative way is not so short as programmatic one — What if you subscribe to 3-4 topic - create/ wrapper (separate static class, that will simply delegate to your service) for each one?

Even if such events will be fired in any case, so, obvious question — what for so complicate things?!

So, until we don't have performance problems with this activity, simply await after component creation and do no complicate without a reason.

GitOrigin-RevId: ca228dfb02519518623fb6665333df466923ef56