Change the meaning of advertiser setting to store ignored stub packages and their versions (PY-31752, PY-31819, PY-34743)
This allows to ignored only incompatible version of stub package, not all of them.
As a result user will be notified about new version of stub package.

The advertiser became more complicated:
on inspection pass it uses cached result and build messages based on it,
after that the advertiser runs background activity to load latest versions for stub packages whose runtime
ones are used in a file.

GitOrigin-RevId: ae9e760206cd864767d5d77ecf648656e80fa500