fix ImmediatePainterTest on Linux
Before commit 9d334ea4, painting in test was performed to ARGB image, automatically switching painting to greyscale antialiasing
(subpixel antialiasing is not supported for transparent targets). This caused inconsistencies in tests, as editor assumed subpixel
antialiasing when calculating font metrics. The commit changed image type to RGB, enabling proper testing of subpixel-antialiased
case. This in turn, uncovered a real problem with current ImmediatePainter implementation, related to the fact, that (at least on
Linux, where FreeType is used to render text) subpixel-antialiased glyph bounds can exceed the bounds defined by glyph's advance by
2 pixels (usually, on macOS and Windows, it's just 1 pixel). This means that the area (background) after newly added glyph should
also be painted by ImmediatePainter for the result to match exactly the output of 'normal' editor painting. To do that we need
to additionally calculate the background color after the newly added character. This is not implemented at the moment. To make the
tests pass nonetheless, they are switched back (now properly set up) to greyscale-antialiased mode.

GitOrigin-RevId: 4ef9f0853edc4c6ba337ed0477b2f28b769c3afd