vcs-ignore: reference completion support for .git/info/exclude for project in git-worktree
For project in git-worktree the affected ignore file root of .git/info/exclude file can be different to .git directory parent. Hence the best way to get the proper affected root is to find Git repository in which .git/info/exclude file belongs to. The root of such repository will be the needed ignore file affected root.

GitOrigin-RevId: 45a62e16da796f868ed32ae272664e6561bbbe65