Another custom Logger.getInstance() inspection
I noticed some invocations like:

    Logger.getInstance("#fqn.of.Class");

And, since the actual FQNs are invariably longer than in my example, I naturally began to wonder if any of these
invocations was misspelled or out-of-date.

So here's an inspection with a quick-fix, which I made by copying the one that rewrites:

    Logger.getInstance("#" + $c$.class.getName())

to:

    Logger.getInstance($c$.class)

and then changing the search template appropriately.

(FYI, I did this directly in the XML, because the inspection settings page just felt too ... cramped and modal.)

I was pleasantly surprised that I didn't need to do anything to turn the FQN text into an expression.

P.S. There are invocations like this in, e.g., Kotlin that would need to be handled separately.

Signed-off-by: Samuel Bronson <naesten@gmail.com>