[vcs-log] introduce openLogTab method in VcsProjectLog
openAnotherLogTab method is not convenient to use since it requires VcsLogManager parameter because VcsLogTabsManager does not keep it in a field. However the method should be called for one specific log manager, the one from VcsProjectLog. So it makes sense to introduce another method to VcsProjectLog so that VcsProjectLog would provide a correct VcsLogManager instance.

GitOrigin-RevId: f728ae034187cfce61304989fecf88adf102144e