git: restore the old Compare with Current dialog
Users miss the ability to view both sides of the comparison in a single view. This problem is not solved yet => the old view is returned for the 2019.3 release.

This commit is a cherry-pick of b432f0f5e5159d745f1569ef68cde1fbba2d5b52 and a revert of 4de067449eabf6a554c8d905d76668225cd5dff6

GitOrigin-RevId: 86db10edf8d3bb2683b65db0b11acea263627f21