PY-38634 Revert registration of PyCharm-specific (vs Python plugin) components
During the recent refactoring a Python plugin only file python-community-plugin-core.xml got included
into the Python plugin which is now bundled with PyCharm. It resulted in using Python plugin specific
components that were designed for the UI of IntelliJ, not PyCharm. This change restores the
registration of PyIdeCommonOptionsFormFactory and moves all includes of python-community-plugin-core.xml
to Python plugin specific files.

GitOrigin-RevId: cd8b6d9b68266ca99f206c005ba73a2e77657950