git: fix loading batch contents for shelving
We can't rely on %objectsize because it counts the size of the content as it is stored in the repository, at least on Windows with standard auto.crlf settings it returns the size of the content as it is stored with \n while the content itself is returned as \r\n.

Instead, let's just split the content by a separator. The hash of the blob is used as an additional protection against possible collision if the separator appears inside the content.

GitOrigin-RevId: 9178c8fa04d2f35385a75afeea805f88482c29cb