[vcs-log] add api for controlling tabs closure by VcsLogTabsWatcher
VcsLogTabsWatcher closes all existing Log tabs on dispose. But sometimes a client that created a tab needs to control the closure. This is the case with the main tab. Before this commit, main tab was handled explicitly by the VcsLogTabsWatcher. This commit removes the hack and introduces a proper api to create such tabs so it could be used by other clients, like branches dashboard.

GitOrigin-RevId: 08dafc14cddc10de363b9005e0cf78038e28a44a