console: try to fix flaky ConsoleViewImplTest.testCompleteLinesWhenMessagesArePrintedConcurrently
When memory is low and/or CPU is busy, the scheduled processing may be invoked later than in asked delay (AWAIT_SAME_STREAM_TEXT_NANO, 10 ms). For example, sometimes scheduled processing is invoked in 100-200 ms instead! Since pending chunks are waiting more than AWAIT_NEW_LINE_NANO in this case, all pending chunk are flushed, even if it's the first scheduled processing and text doesn't end with newline. This leads to mixed stderr/stdout in console.

The fix ensures that pending chunks are not flushed on first scheduled processing when text doesn't end with newline. Not sure though it will help 100%.

GitOrigin-RevId: fdbe2ca5d45de6d12aa97f368cc6a173883fc0f0