PY-38593 Restored missing PyCharm dependencies for running it within IntelliJ
intellij.pycharm.community.main couldn't depend on
intellij.python.community.plugin.main because the latter included a dependency
on IntelliJ Community. Actually this dependency was only used in the saved
run configuration for running IntelliJ Community with the Python plugin. I've
created a separate module for that and removed dependencies on IntelliJ from
*.{pycharm,python}.main modules.

GitOrigin-RevId: c80c439a86b2d552ba40aa54714c637f87470067