git: don't rely on the order of the changes, when loading batch contents for them
We used to rely on the fact that the order of changes, the order of files returned by `git ls-tree`, and the order of contents returned by `git cat-file`, are all the same.

While it seems to be correct that `git cat-file` keeps the order, and we have a protection against failures of this (we check that each content is prefixed by the proper hash), it is not correct for `git ls-tree`: it returns files in a possibly different order than these files were passes as parameters.

(cherry picked from commit d82eb32c0770f88b547821bddfc34e39976325b4)

GitOrigin-RevId: 8929afdea72aaea050fc52d38035eeb318f05db3