[change-reminder] IDEA-216369: Predict VirtualFile instead of FilePath in PredictionProvider
If PredictionProvider predict some FilePaths which are deleted, they won't be showed in UI. So, it will be better to predict some less relative but existed files.
Also, it fixes issue when system predicted `A` for `B` and `A` for `C`, but didn't predict `A` for `B` and `C`. It happened because there were some more relative but deleted files.

(cherry picked from commit e61fc58fb3c5111dad889012ba6551b4470763e6)

GitOrigin-RevId: 334dc632d79b8b027785540bd7025d5e5761a33e