Fix intellij.platform.updater_tests
... and re-enable them.

The problem was that running bazel remotely (either with --config=remote
or via presubmit) would result in test data files having the execute
permission set. The IJ patcher incorporates the execute permission into
its checksum algorithm, so this broke the tests.

This is really a bug in Bazel, but in the meantime
we can work around it.

Fixes: 127523123
Change-Id: I192ef900a0f9ffa933cd9c33cfe5e8ec0e47d490