git: don't cache the error of git version check
Cache the valid version result, but don't cache the error, to handle the situation when Git is installed to the location which used to produce the error.
This can happen, for example, when Git is resolved to /usr/bin/git on Mac, but the Command Line Tools are not installed: we display an error to the user, they install the Tools, and the path to Git becomes correct.

GitOrigin-RevId: 41281fce464d0cf8eeff6595e31418f9891f2ee6