jdks - review feedback - filter JDK entries
several features included into this change:
- filter JDK products
- filter JDK packages
- use predicate to compute default item

Now we check for a "filter" element on JDK product and JDK product/package nodes. The "filter" element has to be an object with "type" field value: `build_number_range`, `and`, `or`, `not`.
The `default` attribute of the JDK product can now be either boolean (e.g. true) or the same filter. It is up to the JSON generator to generate valid `default` attribute predicates for all JDK products listed

IDEA-225308, IDEA-CR-54379

GitOrigin-RevId: 74d1a0aea8e9bbd7e22652cbfc476f1c1ef78c29