Expiration: Fix race with GC when handling already disposed Disposable
On a rare occasion an instance of DisposableExpiration created for an
already disposed Disposable could have been left non-expired. This used
to happen when the GC had already collected a weakly-referenced child
disposable by the time we disposed the instance of
eakReferenceDisposableWrapper referencing the child.

This also fixes
NonBlockingReadActionTest.testSyncExecutionThrowsPCEWhenExpired failing
randomly.

GitOrigin-RevId: 28dfeef0e7d1ef501db0554a5d8069f82603211d