[vcs-log] allow to pause indexing when it is scheduled or in progress
This commit adjusts ResumeIndexingAction so that it could pause indexing as well. Now the action is also enabled when there are unindexed repositories. When used in this mode, indexing is paused for the whole project.

This also means that after pressing "Resume", the action button won't disappear but will change the icon to indicate that indexing was scheduled.

GitOrigin-RevId: f34e7b65ed7dc9842e2db221b4ce80f42ca888eb