git: add a debug log statement indicating that there will be .git/refs file logging further
The dump of .git/refs/ state is written to the debug log, because of a blinking test. However, after another blink there is no logging. This can happen if the whole .git/refs/ directory is removed, and .git/packed-refs as well. So let's add a clear logging statement to indicate that the logging works properly, but there are not files to be logged.

GitOrigin-RevId: db43990b6c5b287e615c1571bc632def3650a524