IDEA-CR-55496: PY-34019 Remove path to the executable from generator's progress messages
It's already been done so for messages on the Java side. The reason is that with
the new UI for progress indicators in the status bar the link can become too
lengthy and, on the whole, it just doesn't have much value: the trailing part
of the path is usually invisible anyway, and it's obvious that it's happening
for the current project interpreter.

GitOrigin-RevId: a363c9acb368cd7e412e7c4f6f712894ed01a950