Revert "IDEA-176615 git: handle unstaged changes during conflicts resolve"
'GitUpdateProcess', 'GitApplyChangesProcess' and others use 'GitConflictResolver.merge'
to check if there are ongoing rebase in the project (and do it before stashing/staging changes).
Thus, we show new dialog whenever there are any unstaged changes in the repository
(and not only when there are an ongoing merge/rebase as it was intended).

This reverts commit 518b754d5eeff9fb6236ca9860659904ffad9c02.

GitOrigin-RevId: 62cb02c49bb2111f384d4fb99af4ff07f889aa3e