ProjectSdksModel - flip the way Sdk#clone() is used in the ProjectSdksModel to allow background updates
The Sdk#clone() is used in the implementation to create the map of <original Sdk -> editable Sdk>.
New behaviour (now with tests) uses the given Sdk instance as editable SDK (was the opposite), while having a cloned instance as the original.
Now it pushes the same editable Sdk instance to the ProjectJdkTable too (now with test)

Overall change is needed to allow SdkDownloader to update the created Sdk instance in the background.
Prior approach used a cloned Sdk instance, so update were easily lost

GitOrigin-RevId: 9185e0f5c980e54aba99aefff11abe9b6981c375