[vcs-log] do not replace main log when adding new log tabs or merging them
Since main tab also displays set filters, its display name could be the same as one of the other tabs. Which means that newly added content can replace it by accident. There is no need for replacement, since the new content is added only when the existing one was not found.

GitOrigin-RevId: 50b032d0cd8c0c3b45734db9ff53207478ad6569