[vcs-log] send validate request on data pack arrival for visible log tabs
When a log tab becomes visible, it needs to be validated. If refresh for some roots was postponed, it should be performed prior to validation so that the tab would use the newest data pack.

Previously, tab was not validated when there were postponed roots, which led to it being stuck in the previous state and in need of manual refresh (could be a possible cause of EA-84824). In most cases, refresh is done in background and no roots are postponed, so the issue did not happen, unless "vcs.log.keep.up.to.date" was disabled or the IDE was in power save mode.

In this commit, validation is added on data pack arrival for all visible tabs.

GitOrigin-RevId: bfae995510018d1553e71e5e7ad0a05c2b86baad