visit functions to determine that __all__ is modified (PY-38322)
Numpy modifies __all__ inside a function. When we detect __all__ modification, we mark it as dynamic and don't use it as a sole source of names.
Not visiting functions breaks resolve of numpy.sin, that is added to it dynamically.

(cherry picked from commit aecfba2cac75a9f32e8180559d6b013ffdc70c47)

GitOrigin-RevId: 7b7bbc30e38f6b65190108ad2639d78f00f5b412