[vcs-log] correctly parse references with parentheses
The list of references for the commit is enclosed in parentheses which are removed as the first step of parsing. But since reference names themselves could contain closing parentheses, the last of them should be removed, not the first one. Removing the first parenthesis results in reference "foo(bar)baz" being parsed as "foo(bar". If there was another reference in the repository named "foo(bar" this could also result in EA-90862.

IDEA-227019

(cherry picked from commit 23db278628bbae734c99fc2229408df0b1a1b158)

GitOrigin-RevId: 13a332a26004edb6ef6a55d511f8cbbe38e18854