jsx: rework solution for WEB-39777, fix WEB-40926
First idea in WEB-39777 was to make "emmet" more priority than completion items. Unfortunately the solution produces a lot of problem: default behavior for "tab" is changed and for users that used to use "tab" for completion workflow is completely broken.
The new solution is to provide "live template" for the most used tags and use default "completion" logic in other cases. As result if user writes something like "div" they can see that the emmet will be used for the completed item.

GitOrigin-RevId: 52c0b7c9937248b3845b812a79d22e3bbb35a143