PY-25586: Support "custom symbols" for test runners.
See issue description for problem.
This is how it is fixed:

When configuration created from context and it has working directory which differs from the one from template configuration, we switch to "custom symbol" mode if test runner supports it.

In this mode we create "Custom" target with "additionalArguments" pointing to certain symbol using framework-specific way.

See ``PyTestConfigurationWithCustomSymbol`` doc

GitOrigin-RevId: 1f5c4e336c19bb9e906642e4b178527c9fb98e37