[workspace model] specify parent directory only in JpsFileEntitySource if file name depends on entity name
Before JpsFileEntitySource stored full path to the configuration file and it caused problems when modules or libraries (in directory-based projects) are renamed. In order to rename corresponding configuration files on save we would need to change entitySource for ModuleEntity and LibraryEntity when their names are changed. This would introduce dependency on serialization internal to the clients code, and it would be hard to properly handle cases when name of the library configuration file is changed to avoid clashes with other libraries which use the same file name because of unsupported name characters' replacement.

Storing only parent directory path in JpsFileEntityStore would make it hard to efficiently reload entities from changed configuration files and properly determine entities which should be saved on change (because all LibraryEntity instances in directory-based project would have equal entitySource). So JpsFileEntitySource.FileInDirectory includes additional fileNameId parameter which is used to distinguish entities loaded from different files.

GitOrigin-RevId: 248cdf4922f581ac9fbf11c50af8de81aa2d03e5