PY-38479 Looking for lib directory inside venv match interpreter paths only if present
They are absent in case of an empty SDK created solely to install package
management tools. In this case we use executable itself as a fallback to
check writability of site-packages. When venv was created with symlinks it
effectively means that we check writability of a system Python executable
and require admin access for no reason. This, in turn, doesn't work with
"pksu" due to PY-32118 that loses the working directory when installing
setuptools.

GitOrigin-RevId: 63a59104b64c317671d59f42ad564df3bd24f31d