[vcs-log] extract diff preview classes and build editor preview outside of MainFrame
Editor diff preview is an external component and is not located on the main panel of the log, so its installation is moved to VcsLogUiImpl. Ordinary (frame) diff preview creation is left inside MainFrame, as it is a part of it, and it can be disabled by a constructor parameter.

Extracted classes FrameDiffPreview and EditorDiffPreview are going to be reused for diff preview in file history in later commits.

GitOrigin-RevId: 4b31fdadb188ffe6b7883ac6a219e702b7a8e1ad