Fix CustomizedActionGroup refresh
Currently, if the underlying ActionGroup that is wrapped by a
CustomizedActionGroup changes after the CustomizedActionGroup
is created, this change is not recognized and the
CustomizedActionGroup will display stale Actions in the
toolbar.

The fix adds property change firings to DefaultActionGroup on
Action add/remove/reset/replacement/etc..., with the
CustomizedActionGroup acting as the property change listener.
Once the property change fires, the CustomizedActionGroup
refreshes its children by querying the DefaultActionGroup
again.