UX-931 deprecate ability to process intention w/o editor.
Intentions can be invoked on a pooled thread with a mock editor, e.g. for intention preview calculating, so erasing editor information is faulty.

For example, some intentions invoke `highlightElements(Collection<PsiElement>, String)` that tries to find selected editor and highlights elements in the original editor instead of the mock.

GitOrigin-RevId: a00850dbcc655ad5fef541c0ea76f5504b9973ae