[vcs-log] do not fire events on scrolling when root names are not displayed
Events are fired on scrolling in order to display a root name only for the first visible row, and leave the column empty for the next rows with the same name. But when names are not shown at all there is no need to do so.

follow-up: ce516eb4b573f40510312382543a4be348dadf0e

GitOrigin-RevId: b1fb171caa34c4c4b037cbf829c51fd627868e75