IDEA-CR-56468: Restore the original text of the editor notification panel for decompiled class file IDEA-227684
The AttachSourcesAction "busy text" is confusing when the action is finished.
There can be cases when AttachSourcesActions are not able to attach source for the class file, so it's better to restore the original text of the panel.

(cherry picked from commit 8005976255adbee8a8aa6f5302b1d1800a5dbe85)

GitOrigin-RevId: 6c3acc2485468023451ad39c608d9815b24cca98