prefer loading a plugin from property over classpath
External plugin developers often use `plugin.path` property to load their plugin in tests,
however to run actual tests with external build tools a classpath entry pointing to the
compiled classes of the plugin is usually passed as well. For some plugins with a complex
artifact, a valid structure is required to properly load, which is provided by the plugin
dist passed with `plugin.path`. When plugin loader overrides it with the one from classpath,
the plugin is loaded but from an incomplete artifact making it fail in cases such as starting
JPS.

GitOrigin-RevId: 64d6fc1760bcc88301c53062f894d60246bfc008