Improve UX of the clone dialog when Git is not installed: IDEA-221692
Check Git executable right from the Clone dialog: on opening, and on receiving the focus. If it is not valid, disable the "Clone" button, and display the error. Depending on the OS, propose different ways to fix the error: on Mac start the default Command Line Tools installer; on Windows download the installer from git-scm.org.

Since now there are 3 ways to display and handle Git executable errors (inline in the clone dialog, modal errors from the Test button in the Settings, and notification balloon in other situations), separate the logic of identifying problem into GitExecutableProblemHandlers, and the UI into ErrorNotifiers.

In the inline UI of the Clone dialog, we need a proper ModalityState to modify the UI after executing the background task, especially to do it immediately after showing the dialog. Thus get the ModalityState when creating dialog components, and pass it to VcsCloneDialogExtensions.

GitOrigin-RevId: e71394d024a4bc6288e40d5594f00a0d873b08d3