IDEA-CR-56721: PY-39764 Restored Python as the primary language for PyCharm
The problem was that an implementation of IdeLanguageCustomization for Python got excluded from PyCharm builds due
to recent refactoring of the Python-related modules. This change brings it back and restores Python as the first
language in the list of languages in the Editor | Code Style settings.

(cherry picked from commit 0e00c712ee1b24a88babec78084b27956a9b53e7)

GitOrigin-RevId: 9fd1794e6c5a41a08e3994b6da3bb6934b76fbca