Fix ShShellcheckExternalAnnotator for Multi Psi File.
If you have a MultiPsiFile (see MultiplePsiFilesPerDocumentFileViewProvider) like different notebooks (Zeppelin, Jupyter) there is a bug which annotates psiElements that belong to the different languages. So you can have a Bash bug in Scala code.

This fix check that element is not an instance of OuterLanguageElement and we need really to mark up this element as an error.

GitOrigin-RevId: 1393110d37dc439a36006eac690edde3e9f35f8b