RUBY-25487 don't attempt to look up ancestors with lowercased names
This caused a cyclic issue and prevented us from caching results properly. When we looked up symbols of a class, we needed to get the symbols contributed from ancestors, but if the include/extend/prepend call was itself passed a method (rather than a constant) then we would try get symbols for `self` again (which in turn needs the symbols from ancestor first... you see the issue).

This fix quickly short-circuits the ancestor lookup if the name does not begin with an uppercase character (constants must have uppercase characters in ruby). The fix also short-circuits the constant lookup if the name is `self` properly (previously it still did the lookup if the self-type didn't match, but in this case we can just quickly return from the method).

GitOrigin-RevId: 29de3982b3d12b1a4859922b6ac049ad53444381