IDEA-229943: Enhance type inference for extracted variables
Intellij IDEA used to travers through all known @FunctionalInterfaces
and checks if their signatures match the examined method reference when
extracting a local variable. It used to skip @FunctionalInterfaces which
have generic return types because type parameters did not get
substituted with their actual values in
FunctionalInterfaceSuggester#composeAcceptableType1 and the types did
not pass the typecheck.

This patch add conversion of type parameters with their actual
representations if there are any.

Signed-off-by: Nikita Eshkeev <kastolom@gmail.com>