IDEA-229943: Enhance type inference for extracted variables
Intellij IDEA used to travers through all known @FunctionalInterfaces
and checks if their signatures match the examined method reference when
extracting a local variable. It used to skip @FunctionalInterfaces which
have generic return types due to inexhaustibleness cases in
TypeConversionUtil#isAssignable().

This patch adds a new case to check if the examined types are type
parameters.

Signed-off-by: Nikita Eshkeev <kastolom@gmail.com>