git: use a single method to request mac command line tools installation
We used to call `/usr/bin/git` if Git is not installed, and `xcode-select --install` if command line tools installation was broken. However, the latter works fine for the first case as well => let's call the latter to DRY.

This also fixes an issue that starting Git installation showed "couldn't start installer". The reason of that was that `/usr/bin/git` invocation failed with an error saying that git is not installed.

A part of IDEA-221692.

GitOrigin-RevId: e608a257c1f328535a302adfbededbf3d401ad61