PersistentHashMap: temporary revert "use nio for indexes" to restore binary compatibility with Kotlin JPS plugin.
Kotlin JPS plugin bundles a "JpsPersistentHashMap" class which is a copy of PersistentHashMap. We didn't detect usages in JPS part of the Kotlin plugin when the breaking API changes were made. Since this class is deprecated in Kotlin source  we will return to using "nio" as soon as Kotlin's sources are cleaned up.

GitOrigin-RevId: 1fbe5eaf33e2e654ddc303dea784fc4193cb8023