Add library libam-instrumentation-data-proto
This is a proto library generated in android.sdktools.ddmlib module
from proto file. The proto is defined in
tools/base/ddmlib/src/main/proto/instrumentation-data.proto.
"am instrument" command outputs test status in this proto message in
serialized bytes and ddmlib needs to parse them. We make a copy of the
proto definition in ddmlib module and generate this library.

Bug: 143299393
Test: exists
Change-Id: Ifa83c1a91afa96a980e9c2d808c6d1acc8b7a800