[hg] Make promthooks extension compatible with python3
but keep compatibility with python2. The correct fix for IDEA-228035.

Mercurial 5.2 stars using python3 when it is available (see https://www.mercurial-scm.org/wiki/Python3), but still could ship with python2. Thus the extension script should work with both python version. To achieve that we need to make sure that correct modules are imported and str type is not used as binary, and is converted to bytes when needed. See https://docs.python.org/3/howto/pyporting.html#text-versus-binary-data

GitOrigin-RevId: 12da5baa1df282678db58faae829bd4fea7c8bc5