IntelliLang: support ignore-pattern option for disabling injection for host or string concatenation
New <ignore-pattern> tag in injection configs determines whether injection to host or concatenation must be ignored.
If at least one place in concatenation is ignored then the entire concatenation must be ignored and language must not be injected.
This option is useful in case the same place can contain different languages or template content that cannot be recognized
as a valid language literal, e.g. SQL with optional Velocity templating.

GitOrigin-RevId: 6fed7e959dc02fc1efc2ed044962caaa3179bfc9