Add filelist to macOS distribution zip
Adds filelist to _codesign directory to indicate the jar files that
contain binaries and dynamic libraries that need to be unpacked and
signed during signing.

The list of jar files comes from the list of warnings generated by
Apple's notarization service when notarizing an incompletely signed copy
of studio. This list must be updated when upstream merges happen or the
distribution will be rejected by the notarization service.

Test: N/A
Bug: 148245411
Change-Id: I62e257d2ae0ab10616d5ae083eed7bec0f3619d4